refactor : 깃허브 OAuth 인증 관련 로직 리팩토링
